rework mesh assets

+++ src/renderer/pass/pbr.rs
@@ -386,7 +386,11 @@ impl PbrRenderPass {
                 render_pass.set_vertex_buffer(0, mesh.vertex_buffer().slice(..));
                 render_pass
                     .set_index_buffer(mesh.index_buffer().slice(..), wgpu::IndexFormat::Uint32);
-                render_pass.draw_indexed(0..mesh.num_indices(), 0, 0..transforms.len() as u32);
+                render_pass.draw_indexed(
+                    0..mesh.num_indices() as u32,
+                    0,
+                    0..transforms.len() as u32,
+                );
             }
 
             queue.submit(Some(encoder.finish()));
